refactor: move sign in methods into lib
